[CLD-7] final refactoring
Commented out all endpoints that are not refactored yet
