+++ backend/api/project_api.py
@@ -6,68 +6,66 @@ import re
 project_api = Blueprint('project_api', __name__)
 
 
-@project_api.route("/projects/all", methods=['GET'])
-def get_projects():
-    id_token = request.args.get('id_token')
-    requestor_email = get_email(id_token)
+# @project_api.route("/projects/all", methods=['GET'])
+# def get_projects():
+#     id_token = request.args.get('id_token')
+#     requestor_email = get_email(id_token)
 
-    invalid_token = check_id_token(id_token, requestor_email)
-    if invalid_token is not None:
-        return make_response(invalid_token), 400
+#     invalid_token = check_id_token(id_token, requestor_email)
+#     if invalid_token is not None:
+#         return make_response(invalid_token), 400
 
-    all_users_col = get_col("users", "users")
-    requestor = all_users_col.find_one({"email": requestor_email})
+#     all_users_col = get_col("users", "users")
+#     requestor = all_users_col.find_one({"email": requestor_email})
 
-    if requestor is None:
-        return user_unauthorised_response()
+#     if requestor is None:
+#         return user_unauthorised_response()
 
-    users_col = get_col("users", "users")
-    requestor = users_col.find_one({'email': requestor_email})
-    names = requestor['projects']
-    response = {'projects': names}
-    return make_response(response), 200
-
-
-@project_api.route('/projects/<project_name>/agreement_score', methods=['GET'])
-def get_agreement_score(project_name):
-    id_token = request.args.get('id_token')
-    requestor_email = get_email(id_token)
-
-    invalid_token = check_id_token(id_token, requestor_email)
-    if invalid_token is not None:
-        return make_response(invalid_token), 400
-
-    user_col = get_col(project_name, "users")
-    requestor = user_col.find_one({'email': requestor_email})
-    if requestor is None:
-        return user_unauthorised_response()
-
-    doc_col = get_col(project_name, "documents")
-
-    all_docs = doc_col.find({})
-    agreed = 0
-    not_agreed = 0
-    for doc in all_docs:
-        if len(doc['user_and_labels']) == 2:
-            label_1 = doc['user_and_labels'][0]['label']
-            label_2 = doc['user_and_labels'][1]['label']
-            if label_1 == label_2:
-                agreed = agreed + 1
-            else:
-                not_agreed = not_agreed + 1
-
-    analysed = agreed + not_agreed
-
-    return_dict = {
-        "agreed_number": agreed,
-        "not_agreed_number": not_agreed,
-        "analysed_number": analysed,
-        "total_number": doc_col.count_documents({})
-    }
-
-    return return_dict, 200
+#     users_col = get_col("users", "users")
+#     requestor = users_col.find_one({'email': requestor_email})
+#     names = requestor['projects']
+#     response = {'projects': names}
+#     return make_response(response), 200
 
 
+# @project_api.route('/projects/<project_name>/agreement_score', methods=['GET'])
+# def get_agreement_score(project_name):
+#     id_token = request.args.get('id_token')
+#     requestor_email = get_email(id_token)
+
+#     invalid_token = check_id_token(id_token, requestor_email)
+#     if invalid_token is not None:
+#         return make_response(invalid_token), 400
+
+#     user_col = get_col(project_name, "users")
+#     requestor = user_col.find_one({'email': requestor_email})
+#     if requestor is None:
+#         return user_unauthorised_response()
+
+#     doc_col = get_col(project_name, "documents")
+
+#     all_docs = doc_col.find({})
+#     agreed = 0
+#     not_agreed = 0
+#     for doc in all_docs:
+#         if len(doc['user_and_labels']) == 2:
+#             label_1 = doc['user_and_labels'][0]['label']
+#             label_2 = doc['user_and_labels'][1]['label']
+#             if label_1 == label_2:
+#                 agreed = agreed + 1
+#             else:
+#                 not_agreed = not_agreed + 1
+
+#     analysed = agreed + not_agreed
+
+#     return_dict = {
+#         "agreed_number": agreed,
+#         "not_agreed_number": not_agreed,
+#         "analysed_number": analysed,
+#         "total_number": doc_col.count_documents({})
+#     }
+
+#     return return_dict, 200
 
 '''
 Create a new project for the current user
@@ -105,34 +103,34 @@ def create_project():
     return "", 204
 
 
-@project_api.route("/projects/<project_name>/delete", methods=['DELETE'])
-def delete_project(project_name):
-    id_token = request.args.get('id_token')
-    requestor_email = get_email(id_token)
+# @project_api.route("/projects/<project_name>/delete", methods=['DELETE'])
+# def delete_project(project_name):
+#     id_token = request.args.get('id_token')
+#     requestor_email = get_email(id_token)
 
-    invalid_token = check_id_token(id_token, requestor_email)
-    if invalid_token is not None:
-        return make_response(invalid_token), 400
+#     invalid_token = check_id_token(id_token, requestor_email)
+#     if invalid_token is not None:
+#         return make_response(invalid_token), 400
 
-    if project_name == "local" or project_name == "users" or project_name == "admin":
-        response = {'message': "Cannot delete that project because it is not a user created project"}
-        return make_response(response), 400
+#     if project_name == "local" or project_name == "users" or project_name == "admin":
+#         response = {'message': "Cannot delete that project because it is not a user created project"}
+#         return make_response(response), 400
 
-    user_col = get_col(project_name, "users")
-    requestor = user_col.find_one({'email': requestor_email, 'isAdmin': True})
-    if requestor is None:
-        return user_unauthorised_response()
+#     user_col = get_col(project_name, "users")
+#     requestor = user_col.find_one({'email': requestor_email, 'isAdmin': True})
+#     if requestor is None:
+#         return user_unauthorised_response()
 
-    my_client = get_db_client()
-    names = my_client.list_database_names()
-    if project_name in names:
-        all_users = user_col.find({})
-        for user in all_users:
-            user_email = user['email']
-            remove_project_from_user(user_email, project_name)
-        my_client.drop_database(project_name)
-    else:
-        response = {'message': "Project does not exist"}
-        return make_response(response), 400
+#     my_client = get_db_client()
+#     names = my_client.list_database_names()
+#     if project_name in names:
+#         all_users = user_col.find({})
+#         for user in all_users:
+#             user_email = user['email']
+#             remove_project_from_user(user_email, project_name)
+#         my_client.drop_database(project_name)
+#     else:
+#         response = {'message': "Project does not exist"}
+#         return make_response(response), 400
 
-    return "", 204
+#     return "", 204
