--- backend/database/model/db.py
@@ -1,6 +1,0 @@
-from flask_mongoengine import MongoEngine
-
-db = MongoEngine()
-
-def initialize_db(app):
-    db.init_app(app)
