--- backend/database/model/label.py
@@ -1,10 +1,0 @@
-import mongoDBInterface
-
-
-class Label:
-
-    def __init__(self, name):
-        self.name = name
-
-
-
